+++ src/main/java/com/cinchapi/runway/Runway.java
@@ -1,33 +1,27 @@
 /*
- * Cinchapi Inc. CONFIDENTIAL
- * Copyright (c) 2017 Cinchapi Inc. All Rights Reserved.
+ * Copyright (c) 2013-2019 Cinchapi Inc.
  *
- * All information contained herein is, and remains the property of Cinchapi.
- * The intellectual and technical concepts contained herein are proprietary to
- * Cinchapi and may be covered by U.S. and Foreign Patents, patents in process,
- * and are protected by trade secret or copyright law. Dissemination of this
- * information or reproduction of this material is strictly forbidden unless
- * prior written permission is obtained from Cinchapi. Access to the source code
- * contained herein is hereby forbidden to anyone except current Cinchapi
- * employees, managers or contractors who have executed Confidentiality and
- * Non-disclosure agreements explicitly covering such access.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * The copyright notice above does not evidence any actual or intended
- * publication or disclosure of this source code, which includes information
- * that is confidential and/or proprietary, and is a trade secret, of Cinchapi.
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * ANY REPRODUCTION, MODIFICATION, DISTRIBUTION, PUBLIC PERFORMANCE, OR PUBLIC
- * DISPLAY OF OR THROUGH USE OF THIS SOURCE CODE WITHOUT THE EXPRESS WRITTEN
- * CONSENT OF COMPANY IS STRICTLY PROHIBITED, AND IN VIOLATION OF APPLICABLE
- * LAWS AND INTERNATIONAL TREATIES. THE RECEIPT OR POSSESSION OF THIS SOURCE
- * CODE AND/OR RELATED INFORMATION DOES NOT CONVEY OR IMPLY ANY RIGHTS TO
- * REPRODUCE, DISCLOSE OR DISTRIBUTE ITS CONTENTS, OR TO MANUFACTURE, USE, OR
- * SELL ANYTHING THAT IT MAY DESCRIBE, IN WHOLE OR IN PART.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package com.cinchapi.runway;
 
+import java.util.Collection;
 import java.util.Set;
 
+import org.reflections.Reflections;
+import org.reflections.scanners.SubTypesScanner;
+
+import com.cinchapi.common.reflect.Reflection;
 import com.cinchapi.concourse.Concourse;
 import com.cinchapi.concourse.ConnectionPool;
 import com.cinchapi.concourse.DuplicateEntryException;
@@ -35,20 +29,42 @@ import com.cinchapi.concourse.lang.BuildableState;
 import com.cinchapi.concourse.lang.Criteria;
 import com.cinchapi.concourse.thrift.Operator;
 import com.cinchapi.concourse.time.Time;
+import com.cinchapi.concourse.util.Logging;
 import com.cinchapi.concourse.util.Strings;
 import com.google.common.base.Throwables;
+import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Iterables;
+import com.google.common.collect.Multimap;
 import com.google.common.collect.Sets;
+
 import gnu.trove.map.TLongObjectMap;
 import gnu.trove.map.hash.TLongObjectHashMap;
 
 /**
+ * {@link Runway} is the ORM controller for Concourse.
+ * <p>
+ * {@link Runway} generally provides methods to retrieve {@link Record} objects.
+ * Subsequent interaction with Records is done using instance methods.
+ * </p>
+ * <p>
+ * If an application has multiple {@link Runway} instances, implicit
+ * {@link Record#save() saving} is disabled in which case the application must
+ * use the {@link #save(Record...)} method provided by this controller.
+ * </p>
  *
- *
- * @author jeff
+ * @author Jeff Nelson
  */
 public final class Runway implements AutoCloseable {
 
+    /**
+     * A mapping from each {@link Record} class to all of its descendants. This
+     * facilitates querying across hierarchies.
+     */
+    private static final Multimap<Class<?>, Class<?>> hierarchies;
+
+    /**
+     * A collection of all the active {@link Runway} instances.
+     */
     private static Set<Runway> instances = Sets.newHashSet();
 
     /**
@@ -58,20 +74,71 @@ public final class Runway implements AutoCloseable {
      */
     private static long METADATA_RECORD = -1;
 
-    public static Runway connect() {
-        return null;
+    static {
+        // NOTE: Scanning the classpath adds startup costs proportional to the
+        // number of classes defined. We do this once at startup to minimize the
+        // effect of the cost.
+        hierarchies = HashMultimap.create();
+        Logging.disable(Reflections.class);
+        Reflections.log = null; // turn off reflection logging
+        Reflections reflection = new Reflections(new SubTypesScanner());
+        reflection.getSubTypesOf(Record.class).forEach(type -> {
+            hierarchies.put(type, type);
+            reflection.getSubTypesOf(type)
+                    .forEach(subType -> hierarchies.put(type, subType));
+        });
     }
 
+    /**
+     * Return a {@link Runway} instance that is connected to Concourse using the
+     * default connection parameters.
+     * 
+     * @return a {@link Runway} instance
+     */
+    public static Runway connect() {
+        return new Runway(ConnectionPool.newCachedConnectionPool());
+    }
+
+    /**
+     * Return a {@link Runway} instance that is connected to Concourse using the
+     * provided connection parameters.
+     * 
+     * @param host
+     * @param port
+     * @param username
+     * @param password
+     * @return a {@link Runway} instance
+     */
     public static Runway connect(String host, int port, String username,
             String password) {
         return connect(host, port, username, password, "");
     }
 
+    /**
+     * Return a {@link Runway} instance that is connected to Concourse using the
+     * provided connection parameters.
+     * 
+     * @param host
+     * @param port
+     * @param username
+     * @param password
+     * @param environment
+     * @return a {@link Runway} instance
+     */
     public static Runway connect(String host, int port, String username,
             String password, String environment) {
-        return new Runway(host, port, username, password, environment);
+        return new Runway(ConnectionPool.newCachedConnectionPool(host, port,
+                username, password, environment));
     }
 
+    /**
+     * Utility method do ensure that the {@code criteria} is limited to querying
+     * objects that belong to a specific {@code clazz}.
+     * 
+     * @param criteria
+     * @param clazz
+     * @return the updated {@code criteria}
+     */
     private static <T> Criteria ensureClassSpecificCriteria(Criteria criteria,
             Class<T> clazz) {
         return Criteria.where().key(Record.SECTION_KEY)
@@ -95,16 +162,10 @@ public final class Runway implements AutoCloseable {
     /**
      * Construct a new instance.
      * 
-     * @param host
-     * @param port
-     * @param username
-     * @param password
-     * @param environment
+     * @param connections a Concourse {@link ConnectionPool}
      */
-    private Runway(String host, int port, String username, String password,
-            String environment) {
-        this.connections = ConnectionPool.newCachedConnectionPool(host, port,
-                username, password, environment);
+    private Runway(ConnectionPool connections) {
+        this.connections = connections;
         instances.add(this);
         if(instances.size() > 1) {
             Record.PINNED_RUNWAY_INSTANCE = null;
@@ -168,6 +229,86 @@ public final class Runway implements AutoCloseable {
     }
 
     /**
+     * Execute the {@link #find(Class, BuildableState)} query for {@code clazz}
+     * and all of its descendants.
+     * 
+     * @param clazz
+     * @param criteria
+     * @return the matching records
+     */
+    public <T extends Record> Set<T> findAny(Class<T> clazz,
+            BuildableState criteria) {
+        return findAny(clazz, criteria.build());
+    }
+
+    /**
+     * Execute the {@link #find(Class, Criteria)} query for {@code clazz} and
+     * all of its descendants.
+     * 
+     * @param clazz
+     * @param criteria
+     * @return the matching records
+     */
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+    public <T extends Record> Set<T> findAny(Class<T> clazz,
+            Criteria criteria) {
+        Collection<Class<?>> hierarchy = hierarchies.get(clazz);
+        Set<T> found = Sets.newLinkedHashSet();
+        for (Class cls : hierarchy) {
+            found.addAll(find(cls, criteria));
+        }
+        return found;
+    }
+
+    /**
+     * Execute the {@link #findUnique(Class, BuildableState)} query for
+     * {@code clazz} and all of its descendants.
+     * 
+     * @param clazz
+     * @param criteria
+     * @return the one matching record
+     */
+    public <T extends Record> T findAnyUnique(Class<T> clazz,
+            BuildableState criteria) {
+        return findAnyUnique(clazz, criteria.build());
+    }
+
+    /**
+     * Execute the {@link #findUnique(Class, Criteria)} query for {@code clazz}
+     * and
+     * all of its descendants.
+     * 
+     * @param clazz
+     * @param criteria
+     * @return the one matching record
+     */
+    @SuppressWarnings({ "unchecked", "rawtypes" })
+    public <T extends Record> T findAnyUnique(Class<T> clazz,
+            Criteria criteria) {
+        Collection<Class<?>> hierarchy = hierarchies.get(clazz);
+        Set<T> found = Sets.newLinkedHashSet();
+        for (Class cls : hierarchy) {
+            T $found = (T) findUnique(cls, criteria);
+            if($found != null) {
+                found.add($found);
+            }
+        }
+        if(found.size() == 0) {
+            return null;
+        }
+        else if(found.size() == 1) {
+            return found.iterator().next();
+        }
+        else {
+            throw new DuplicateEntryException(
+                    new com.cinchapi.concourse.thrift.DuplicateEntryException(
+                            Strings.format(
+                                    "There are more than one records that match {} in the hierarchy of {}",
+                                    criteria, clazz)));
+        }
+    }
+
+    /**
      * Find the one record of type {@code clazz} that matches the
      * {@code criteria}. If more than one record matches, throw a
      * {@link DuplicateEntryException}.
@@ -176,10 +317,26 @@ public final class Runway implements AutoCloseable {
      * @param criteria
      * @return the one matching record
      * @throws DuplicateEntryException
+     * @deprecated use {@link #findUnique(Class, BuildableState)}
      */
     public <T extends Record> T findOne(Class<T> clazz,
             BuildableState criteria) {
-        return findOne(clazz, criteria.build());
+        return findUnique(clazz, criteria);
+    }
+
+    /**
+     * Find the one record of type {@code clazz} that matches the
+     * {@code criteria}. If more than one record matches, throw a
+     * {@link DuplicateEntryException}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @return the one matching record
+     * @throws DuplicateEntryException
+     * @deprecated use {@link, #findUnique(Class, Criteria)}
+     */
+    public <T extends Record> T findOne(Class<T> clazz, Criteria criteria) {
+        return findUnique(clazz, criteria);
     }
 
     /**
@@ -192,7 +349,22 @@ public final class Runway implements AutoCloseable {
      * @return the one matching record
      * @throws DuplicateEntryException
      */
-    public <T extends Record> T findOne(Class<T> clazz, Criteria criteria) {
+    public <T extends Record> T findUnique(Class<T> clazz,
+            BuildableState criteria) {
+        return findUnique(clazz, criteria.build());
+    }
+
+    /**
+     * Find the one record of type {@code clazz} that matches the
+     * {@code criteria}. If more than one record matches, throw a
+     * {@link DuplicateEntryException}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @return the one matching record
+     * @throws DuplicateEntryException
+     */
+    public <T extends Record> T findUnique(Class<T> clazz, Criteria criteria) {
         Concourse concourse = connections.request();
         try {
             criteria = ensureClassSpecificCriteria(criteria, clazz);
@@ -262,8 +434,48 @@ public final class Runway implements AutoCloseable {
      * @return the existing Record
      */
     public <T extends Record> T load(Class<T> clazz, long id) {
+        if(hierarchies.get(clazz).size() > 1) {
+            // The provided clazz has descendants, so it is possible that the
+            // Record with the #id is actually a member of a subclass
+            Concourse connection = connections.request();
+            try {
+                String section = connection.get(Record.SECTION_KEY, id);
+                if(section != null) {
+                    clazz = Reflection.getClassCasted(section);
+                }
+            }
+            finally {
+                connections.release(connection);
+            }
+        }
         return load(clazz, id, new TLongObjectHashMap<Record>());
     }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} or any of its descendants.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+    public <T extends Record> Set<T> loadAny(Class<T> clazz) {
+        Collection<Class<?>> hierarchy = hierarchies.get(clazz);
+        Set<T> loaded = Sets.newLinkedHashSet();
+        for (Class cls : hierarchy) {
+            loaded.addAll(load(cls));
+        }
+        return loaded;
+    }
+
+    // TODO: what about loading a specific record using a parent class?
 
     /**
      * Save all the changes in all of the {@code records} using a single ACID